feat(pix-structure-switcher): refactored pix select by creating a pix select list component
