+++ addon/components/pix-select-list.hbs
@@ -1,0 +1,86 @@
+<ul role="listbox" id={{@listId}} class="pix-select_list" ...attributes>
+  <li
+    class="pix-select-list-category__option{{unless
+        @value
+        ' pix-select-list-category__option--selected'
+      }}{{unless @displayDefaultOption ' pix-select-list-category__option--hidden'}}"
+    role="option"
+    tabindex={{if this.isDefaultOptionHidden "-1" "0"}}
+    aria-selected={{if @value "false" "true"}}
+    {{on "click" (fn @onChange @defaultOption)}}
+    {{on-enter-action (fn @onChange @defaultOption)}}
+    {{on-space-action (fn @onChange @defaultOption)}}
+  >
+    {{@defaultOptionValue}}
+  </li>
+  {{#if this.results}}
+    {{#if this.displayCategory}}
+      {{#each this.results as |element index|}}
+        <ul
+          class="pix-select-list-category"
+          role="group"
+          aria-labelledby={{if this.displayCategory (concat "cat-" @selectId "-" index)}}
+        >
+          {{#if this.displayCategory}}
+            <li
+              class="pix-select-list-category__name"
+              role="presentation"
+              id={{concat "cat-" @selectId "-" index}}
+              title={{element.label}}
+            >
+              {{element.category}}
+            </li>
+          {{/if}}
+
+          {{#each element.options as |option|}}
+            {{! template-lint-disable require-context-role }}
+            {{!https://www.w3.org/WAI/ARIA/apg/example-index/listbox/listbox-grouped.html}}
+            <li
+              class="pix-select-list-category__option{{if
+                  (eq option.value @value)
+                  ' pix-select-list-category__option--selected'
+                }}"
+              role="option"
+              tabindex={{if @isExpanded "0" "-1"}}
+              title={{option.label}}
+              aria-selected={{if (eq option.value @value) "true" "false"}}
+              {{on "click" (fn @onChange option)}}
+              {{on-enter-action (fn @onChange option)}}
+              {{on-space-action (fn @onChange option)}}
+            >
+              {{option.label}}
+
+              {{#if (eq option.value @value)}}
+                <PixIcon @name="check" role="presentation" />
+              {{/if}}
+            </li>
+          {{/each}}
+        </ul>
+      {{/each}}
+    {{else}}
+      {{#each this.results as |option|}}
+        <li
+          class="pix-select-list-category__option{{if
+              (eq option.value @value)
+              ' pix-select-list-category__option--selected'
+            }}"
+          role="option"
+          tabindex={{if @isExpanded "0" "-1"}}
+          aria-selected={{if (eq option.value @value) "true" "false"}}
+          title={{option.label}}
+          {{on "click" (fn @onChange option)}}
+          {{on-enter-action (fn @onChange option)}}
+          {{on-space-action (fn @onChange @defaultOption)}}
+        >
+          {{option.label}}
+
+          {{#if (eq option.value @value)}}
+            <PixIcon @name="check" role="presentation" />
+          {{/if}}
+        </li>
+      {{/each}}
+    {{/if}}
+  {{else}}
+    <li class="pix-select-list__empty-search-message">{{@emptySearchMessage}}</li>
+  {{/if}}
+</ul>
