+++ addon/styles/addon.scss
@@ -15,6 +15,7 @@
 @import 'pix-progress-gauge';
 @import 'pix-return-to';
 @import 'pix-select';
+@import 'pix-select-list';
 @import 'pix-stars';
 @import 'pix-tag';
 @import 'pix-textarea';
