+++ app/components/pix-select-list.js
@@ -1,0 +1,1 @@
+export { default } from '@1024pix/pix-ui/components/pix-select-list';
